refactor: add $padding arg to Base64\encode and decode
In Base64 namespace:

For both encode and url_encode, add bool $padding arg to tell if we have
or not padding at the end. By default encode is set to true and
url_encode to false.

For both decode and url_decode, add bool $explicitPadding arg to tell if
we check or not padding at the end. By default decode is set to true and
url_decode to false.